Add routes with react-router-dom
