+++ src/components/Home.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const Home = () => {
+  const name = 'ddd';
+  return <div>{name}</div>;
+};
+
+export default Home;
