+++ src/components/Quotes.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const Quotes = () => {
+  const name = 'ddd';
+  return <div>{name}</div>;
+};
+
+export default Quotes;
